cli/autoinstall: Move CLI functions to local class and replace testing_data_generator with proper user_create_user() function call
